Fixed str #1147
git-svn-id: file:///fltk/svn/trunk@4939 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/GSave.cxx
@@ -81,7 +81,7 @@ GSave::GSave() {
 }
 
 GSave::~GSave() {
-  unsigned v = (unsigned)data[3];
+  unsigned v = (unsigned) (unsigned long)data[3];
   fl_clip_w = v >> 16;
   fl_clip_h = v & 0xffff;
 #if USE_X11
